[WIP] Add Warp Sync
This patch adds the basis of Warp Sync to Mana.

Warp Sync is documented [here](https://wiki.parity.io/Warp-Sync), [here](https://wiki.parity.io/Warp-Sync-Snapshot-Format.html), and, as always, in the source code of implementations that isn't written down anywhere else.

The core thesis to warp sync is that we can give block headers, transactions, receipts and "rich account" data, which, in sum, let us know the state of the blockchain at block N (without knowing about the state for blocks 0-N). We can then sync blocks 0-N in the background while being an up-to-date client.

This is accomplished through Manifests, BlockChunks (headers, transactions and receipts) and StateChunks (account data). The data is sent over-the-wire with Snappy compression to reduce overhead.

This patch has a variety of TODOs, but the biggest is once we receive block chunks, how do we process them so that we're "up-to-date". The process is mostly "store Key-Values to the DB," which then we need to make sure we do it correctly (and without excessive processing). Also, if we queue up a huge amount of trie changes, we should make sure we parallelize that process as much as possible. That said, we do update our sync module to be in one of two states, either warp sync (which requests block/state chunks for warp) or non-warp sync (which processes blocks via the block queue). It appears that a client must handle both these states.

By and large, this patch gets us to processing Warp correctly with Parity and takes us one step closer to a legit client.

* Apply suggestions from code review
* Get block chunk sync fully working